Move server into subpackage.
